changed the wrong page earlier, now it must be fine, also populated stars to check paging
